Separacion de cada proxy + extraccion de la creacion de una cardData en un funcion aparte
